An abstract search class and basic document search class
